Enhance app functionality and code quality
- Switched `UpdateMode` to `Background` in `Deno.csproj`.
- Updated `ApplicationRevision` to `91` and added desktop shortcut.
- Increased `WarningLevel` to `5` for stricter code checks.
- Refactored `CurrencyService.cs`:
  - Removed unused `using` directives.
  - Changed `InitialAmount` type from `decimal` to `float`.
  - Added support for suspended bills, including an HTML template.
  - Simplified `PostToApi` by removing redundant authentication logic.
- Adjusted `HomePage.xaml` and `HomePage.xaml.cs` for `float` parsing.
- Improved error messaging in `LoginWindow.xaml.cs`.
- General code cleanup and UI enhancements for better usability.

+++ Views/HomePage.xaml.cs
@@ -48,7 +48,7 @@ namespace Deno.Views
             _currencyService = currencyService;
             _username = username;
 
-            _currencyService.InitialAmount = decimal.TryParse(GlobalStateService.Instance.OpeningAmount, out var openingAmount) ? openingAmount : 0m;
+            _currencyService.InitialAmount = float.TryParse(GlobalStateService.Instance.OpeningAmount, out var openingAmount) ? openingAmount : 0;
 
             WelcomeText = $"Welcome, {username}!";
             DataContext = _currencyService;
